Añadido alert al delete

+++ public/js/main.js
@@ -1,11 +1,13 @@
 deleteComic = async (id) => {
-  await fetch("http://localhost:3200/comics/" + id, {
-    method: "DELETE",
+  if (confirm("DO YOU REALLY WANT TO DELETE THIS COMIC?")) {
+    await fetch("http://localhost:3200/comics/" + id, {
+      method: "DELETE",
 
-    headers: {
-      "Content-type": "application/json; charset=UTF-8",
-    },
-  });
+      headers: {
+        "Content-type": "application/json; charset=UTF-8",
+      },
+    });
 
-  window.location.href = "/comics";
+    window.location.href = "/comics";
+  }
 };
